+++ hospital/DAO/LastIdGetter.cs
@@ -6,32 +6,32 @@ namespace hospital.DAO
     {
 
         protected string GetLastID;
-        protected uint GetLastId(MySqlConnection connection, MySqlTransaction transaction)
+        protected long GetLastId(MySqlConnection connection, MySqlTransaction transaction)
         {
-            uint id = 0;
+            long id = 0;
             using (MySqlCommand command = new MySqlCommand(GetLastID, connection))
             {
                 command.Transaction = transaction;
                 var result = command.ExecuteScalar();
                 if (result != null && result != DBNull.Value)
                 {
-                    id = Convert.ToUInt32(result);
+                    id = Convert.ToInt64(result);
                 }
             }
 
             return id;
         }
 
-        protected uint GetLastId(MySqlConnection connection, MySqlTransaction transaction,string GetLastIdQuery)
+        protected long GetLastId(MySqlConnection connection, MySqlTransaction transaction,string GetLastIdQuery)
         {
-            uint id = 0;
+            long id = 0;
             using (MySqlCommand command = new MySqlCommand(GetLastIdQuery, connection))
             {
                 command.Transaction = transaction;
                 var result = command.ExecuteScalar();
                 if (result != null && result != DBNull.Value)
                 {
-                    id = Convert.ToUInt32(result);
+                    id = Convert.ToInt64(result);
                 }
             }
