+++ hospital/Entities/Drug.cs
@@ -2,12 +2,12 @@
 {
     public class Drug
     {
-        public uint Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Instruction { get; set; }
         public DateTime ExpirationDate { get; set; }
         public bool IsSelected {  get; set; }
-        public Drug(uint id, string name, string instruction, DateTime expirationDate)
+        public Drug(long id, string name, string instruction, DateTime expirationDate)
         {
             Id = id;
             Name = name;
